Declaring server as local variable in main
Replaced luv_server_create with luv_server_init which accepts a
luv_server_t as the first argument. This allows us to pass in a
variable to be used as the server.

There is no need to malloc something that can be declared in main. Since the
event loop keeps main from returning we can be assured that the
server variable will continue to live on the stack.

Now we don't have to worry about freeing the server variable. TODO:
Call or remove luv_server_destroy. Use it or lose it.