Using strlen to count buf len for luv_server_send
snprintf's return value can't be counted on to tell us the length of
the printed buffer (duh!), so we call strlen before calling
luv_server_send. This is still better than calling strlen inside
of luv_server_send because it avoids calling strlen redundantly as in
the case of luv_server_broadcast where the same message is being
sent out to different clients.